toggle dbconfig.tablename at one place only
- selecting which database table to work on can be done via the dbconfig object.
This commit introduces a variable that holds the dbconfig.TABLENAME as a constant to be used in the code.

The result is a single place in the code to set which database table to use during execution. Avoiding mistakes that have happend due to inconsitent changes of that propertie made in various DIFFERENT lines of code that should have all used the same value.